IDE Cleanup (#3323)
* Concat if statements with return statement

IDE-Automated

* Remove redundant boxing/unboxing

IDE-Assisted

* Replace indexOf with contains

IDE-Performed

* Replace loops with foreach if possible

IDE-Assisted

* Use decimal instead of octal 0

IDE-Performed

* Simplify arithmetic expression

IDE-Performed

* Collapse identical catch blocks

* Replace lambdas with method reference

IDE-Performed

* Use Comparator combinators

IDE-Performed

* Replace statement lambdas with expression lambdas

IDE-Performed

* Replace Collections#sort with List#sort

IDE-Performed

* Fix potential performance issues

* Javadoc -> block comment

IDE-Performed

* Remove final modifier on static methods

IDE-Performed

* Turn fields into local variables if possible

IDE-Assisted

* Fix checkstyle violations

* Fix minor issues

* Fix some further formatting issues

* Fix remaining formatting issues

+++ game-core/src/main/java/games/strategy/engine/data/export/GameDataExporter.java
@@ -181,7 +181,7 @@ public class GameDataExporter {
     }
     if (prop.getClass().equals(ColorProperty.class)) {
       typeString = "            <color/>\n";
-      value = "0x" + Integer.toHexString((((Integer) prop.getValue()).intValue())).toUpperCase();
+      value = "0x" + Integer.toHexString(((Integer) prop.getValue())).toUpperCase();
     }
     if (prop.getClass().equals(ComboProperty.class)) {
       final Field listField;
